+++ app.js
@@ -5,6 +5,8 @@ const cookieParser = require('cookie-parser');
 const bodyParser = require('body-parser');
 
 const routes = require('./routes/index');
+const objectives = require('./routes/objectives');
+const progressUpdates = require('./routes/progress-updates');
 
 const app = express();
 
@@ -21,6 +23,8 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', routes);
+app.use('/objectives', objectives);
+app.use('/progress-updates', progressUpdates);
 
 // catch 404 and forward to error handler
 app.use((req, res, next) => {
