+++ routes/index.js
@@ -1,13 +1,5 @@
 const express = require('express');
 const bodyParser = require('body-parser');
-const models = require('../server/models/index');
-const {
-  validateData,
-  objectiveDataSchema,
-  updateDataSchema,
-  updateProgressUpdateSchema,
-  progressUpdateSchema,
-} = require('./validation');
 
 const router = express.Router();
 router.use(bodyParser.json());
@@ -15,184 +7,6 @@ router.use(bodyParser.urlencoded({ extended: false }));
 
 router.get('/', (req, res) => {
   res.render('index');
-});
-
-// CREATE OBJECTIVE
-router.post('/objectives', (req, res) => {
-  const {
-    body: {
-      title,
-      type,
-      totalPagesVideos,
-      timeAllocated,
-    },
-  } = req;
-  const validatedRequestData = validateData(req.body, objectiveDataSchema);
-  return !validatedRequestData.isValid ?
-    res.status(400).send(`ERROR 400: ${ validatedRequestData.errorMessage }`) :
-    models.Objective.findOrCreate({
-      where: {
-        dateCreated: new Date(),
-        title,
-        type,
-        totalPagesVideos,
-        timeAllocated,
-        completed: false,
-      },
-    }).then((response) => {
-      const data = response[0].dataValues;
-      res.status(200).send(data);
-    });
-});
-
-// READ ALL OBJECTIVES
-router.get('/objectives', (req, res) => {
-  models.Objective.findAll()
-    .then(objectives => (res.status(200).send(objectives)))
-    .catch(error => error);
-});
-
-// READ A SINGLE OBJECTIVE
-router.get('/objectives/:id', (req, res) => {
-  const { params: { id } } = req;
-  models.Objective.findById(id)
-    .then(objective => (res.status(200).send(objective)))
-    .catch(error => error);
-});
-
-// UPDATE SINGLE OBJECTIVE
-router.patch('/objectives/:id', (req, res) => {
-  const { body, params } = req;
-  const { title, type, totalPagesVideos, timeAllocated } = body || null;
-  const validatedUpdateData = validateData(body, updateDataSchema);
-  return !validatedUpdateData.isValid ?
-    res.status(400).send(`ERROR 400: ${ validatedUpdateData.errorMessage }`) :
-    models.Objective.update({
-      title,
-      type,
-      totalPagesVideos,
-      timeAllocated,
-    }, {
-      where: { id: params.id },
-      returning: true,
-    })
-      .then(objectives => (objectives[1].length > 0 ?
-        res.status(200).send(objectives[1][0]) :
-        res.status(404).send(`ERROR 404: An objective with the ID of ${ params.id } does not exist`)
-      ))
-      .catch(error => error);
-});
-
-// DELETE SINGLE OBJECTIVE
-router.delete('/objectives/:id', (req, res) => {
-  const { params: { id } } = req;
-  models.Objective.destroy({
-    where: { id },
-  })
-    .then(response => (
-      response === 1 ?
-        res.status(200).send(`Objective ID: ${ id } has been deleted successfully.`) :
-        res.status(404).send(`ERROR 404: An objective with the ID: ${ id } has not been found`)
-    ))
-    .catch(error => error);
-});
-
-// ADD PROGRESS UPDATE
-router.post('/objectives/:id/progress-updates', (req, res) => {
-  const {
-    body: {
-      objectiveId,
-      pageVideoNumReached,
-      learningSummary,
-    },
-    params,
-  } = req;
-  const objectiveIdsMatch = objectiveId === parseInt(params.id, 10);
-  const nonMatchingIdsError =
-    `The objectiveId (${ objectiveId }) and the id provided in the request (${ params.id }) do not match`;
-  const validatedRequestData = validateData(req.body, progressUpdateSchema);
-  return validatedRequestData.isValid && objectiveIdsMatch ?
-    models.Objective.findById(params.id)
-      .then(objective => (objective ?
-        models.ProgressUpdate.findOrCreate({
-          where: {
-            dateCreated: new Date(),
-            objectiveId,
-            pageVideoNumReached,
-            learningSummary,
-          },
-        })
-          .then((response) => {
-            const data = response[0].dataValues;
-            res.status(200).send(data);
-          }) :
-        res.status(404).send(`ERROR 404: The objective with an ID of ${ params.id } does not exist`)
-      ))
-      .catch(error => error) :
-    res.status(400).send(
-      `ERROR 400: ${ validatedRequestData.errorMessage || nonMatchingIdsError }` // eslint-disable-line comma-dangle
-    );
-});
-
-// READING ALL PROGRESS UPDATES FOR A SINGLE OBJECTIVE
-router.get('/objectives/:id/progress-updates', (req, res) => {
-  const { params: { id } } = req;
-  models.ProgressUpdate.findAll({
-    where: { objectiveId: id },
-  })
-    .then(progressUpdates => (res.status(200).send(progressUpdates)))
-    .catch(error => error);
-});
-
-// READING AN SINGLE PROGRESS UPDATE
-router.get('/progress-updates/:id', (req, res) => {
-  const { params: { id } } = req;
-  models.ProgressUpdate.findById(id)
-    .then(progressUpdate => (res.status(200).send(progressUpdate)))
-    .catch(error => error);
-});
-
-// READ ALL PROGRESS UPDATES
-router.get('/progress-updates', (req, res) => {
-  models.ProgressUpdate.findAll()
-    .then(progressUpdates => (res.status(200).send(progressUpdates)))
-    .catch(error => error);
-});
-
-// UPDATE SINGLE PROGRESS UPDATE
-router.patch('/progress-updates/:id', (req, res) => {
-  const { body, params } = req;
-  const { pageVideoNumReached, learningSummary, objectiveId } = body || null;
-  const objectiveIdError = 'The objectiveId cannot be updated once added';
-  const validatedUpdateData = validateData(body, updateProgressUpdateSchema);
-  return !validatedUpdateData.isValid || objectiveId ?
-    res.status(400).send(`ERROR 400: ${ validatedUpdateData.errorMessage || objectiveIdError }`) :
-    models.ProgressUpdate.update({
-      pageVideoNumReached,
-      learningSummary,
-    }, {
-      where: { id: params.id },
-      returning: true,
-    })
-      .then(objectives => (objectives[1].length > 0 ?
-        res.status(200).send(objectives[1][0]) :
-        res.status(404).send(`ERROR 404: A progress update with the ID of ${ params.id } does not exist`)
-      ))
-      .catch(error => error);
-});
-
-// DELETE SINGLE PROGRESS UPDATE
-router.delete('/progress-updates/:id', (req, res) => {
-  const { params: { id } } = req;
-  models.ProgressUpdate.destroy({
-    where: { id },
-  })
-    .then(response => (
-      response === 1 ?
-        res.status(200).send(`Progress update ID: ${ id } has been deleted successfully.`) :
-        res.status(404).send(`ERROR 404: A progress update with the ID: ${ id } has not been found`)
-    ))
-    .catch(error => error);
 });
 
 module.exports = router;
